refactor v25 code and add more tests
