+++ client.go
@@ -18,7 +18,6 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials"
-	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
@@ -28,17 +27,7 @@ import (
 )
 
 const (
-	cloudPort    = "443"
-	dgraphScheme = "dgraph"
-	// optional parameter for providing a Dgraph Cloud API key
-	cloudAPIKeyParam = "apikey"
-	// optional parameter for providing an access token
-	bearerTokenParam = "bearertoken"
-	// optional parameter for providing a Dgraph SSL mode
-	sslModeParam    = "sslmode"
-	sslModeDisable  = "disable"
-	sslModeRequire  = "require"
-	sslModeVerifyCA = "verify-ca"
+	cloudPort = "443"
 )
 
 // Dgraph is a transaction-aware client to a Dgraph cluster.
@@ -63,82 +52,6 @@ func (a *authCreds) GetRequestMetadata(ctx context.Context, uri ...string) (
 
 func (a *authCreds) RequireTransportSecurity() bool {
 	return true
-}
-
-// Open creates a new Dgraph client by parsing a connection string of the form:
-// dgraph://<optional-login>:<optional-password>@<host>:<port>?<optional-params>
-// For example `dgraph://localhost:9080?sslmode=require`
-//
-// Parameters:
-// - apikey: a Dgraph Cloud API key for authentication
-// - bearertoken: a token for bearer authentication
-// - sslmode: SSL connection mode (options: disable, require, verify-ca)
-//   - disable: No TLS (default)
-//   - require: Use TLS but skip certificate verification
-//   - verify-ca: Use TLS and verify the certificate against system CA
-//
-// If credentials are provided, Open connects to the gRPC endpoint and authenticates the user.
-// An error can be returned if the Dgraph cluster is not yet ready to accept requests--the text
-// of the error in this case will contain the string "Please retry".
-func Open(connStr string) (*Dgraph, error) {
-	u, err := url.Parse(connStr)
-	if err != nil {
-		return nil, fmt.Errorf("invalid connection string: %w", err)
-	}
-
-	if u.Scheme != dgraphScheme {
-		return nil, fmt.Errorf("invalid scheme: must start with %s://", dgraphScheme)
-	}
-
-	opts := []ClientOption{}
-
-	apiKey := u.Query().Get(cloudAPIKeyParam)
-	bearerToken := u.Query().Get(bearerTokenParam)
-	sslMode := u.Query().Get(sslModeParam)
-
-	if apiKey != "" && bearerToken != "" {
-		return nil, errors.New("invalid connection string: both apikey and bearertoken cannot be provided")
-	}
-
-	if apiKey != "" {
-		opts = append(opts, WithDgraphAPIKey(apiKey))
-	}
-
-	if bearerToken != "" {
-		opts = append(opts, WithBearerToken(bearerToken))
-	}
-
-	if sslMode == "" {
-		sslMode = sslModeDisable
-	}
-	switch sslMode {
-	case sslModeDisable:
-		opts = append(opts, WithGrpcOption(grpc.WithTransportCredentials(insecure.NewCredentials())))
-	case sslModeRequire:
-		opts = append(opts, WithSkipTLSVerify())
-	case sslModeVerifyCA:
-		opts = append(opts, WithSystemCertPool())
-	default:
-		return nil, fmt.Errorf("invalid SSL mode: %s (must be one of %s, %s, %s)", sslMode, sslModeDisable, sslModeRequire, sslModeVerifyCA)
-	}
-
-	if u.User != nil {
-		username := u.User.Username()
-		password, _ := u.User.Password()
-		if username == "" || password == "" {
-			return nil, errors.New("invalid connection string: both username and password must be provided")
-		}
-		opts = append(opts, WithACLCreds(username, password))
-	}
-
-	return NewClient(u.Host, opts...)
-}
-
-// Close shutdown down all the connections to the Dgraph Cluster.
-func (d *Dgraph) Close() {
-	for _, conn := range d.conns {
-		_ = conn.Close()
-	}
 }
 
 // NewDgraphClient creates a new Dgraph (client) for interacting with Alphas.
@@ -260,9 +173,13 @@ func (d *Dgraph) LoginIntoNamespace(ctx context.Context,
 // Deprecated: use DropAllNamespaces, DropAll, DropData, DropPredicate, DropType, SetSchema instead.
 func (d *Dgraph) Alter(ctx context.Context, op *api.Operation) error {
 	dc := d.anyClient()
-	_, err := doWithRetryLogin(ctx, d, func() (*api.Payload, error) {
-		return dc.Alter(d.getContext(ctx), op)
-	})
+	_, err := dc.Alter(d.getContext(ctx), op)
+	if isJwtExpired(err) {
+		if err := d.retryLogin(ctx); err != nil {
+			return err
+		}
+		_, err = dc.Alter(d.getContext(ctx), op)
+	}
 	return err
 }
